WIP: replace divs and spans & use classNames
