--- src/components/panes/index.ts
@@ -1,1 +1,0 @@
-export * from './DesignedForHumansPane';
